#881 move types folder
