+++ digiwf-apps/packages/components/digiwf-dms-input/src/middleware/dmsMiddleware.ts
@@ -1,9 +1,16 @@
 import { Configuration, FetchUtils, Metadata } from "@muenchen/digiwf-dms-api-internal";
-import {Objectclass} from "@/types";
 import {
   getMetadataFromDmsservice
 } from "@/apiClient/dmsApiCalls";
 
+export enum Objectclass {
+  Sachakte ="Sachakte",
+  Vorgang = "Vorgang",
+  Eingang = "Eingang",
+  Ausgang = "Ausgang",
+  Intern = "Intern",
+  Schriftstueck = "Schriftstueck",
+}
 
 export const getMetadata = async (ojectclass: Objectclass, coo: string, apiEndpoint: string): Promise<Metadata> => {
   const dmsAxiosConfig = axiosConfig(apiEndpoint);
