+++ digiwf-apps/packages/components/digiwf-dms-input/types/index.d.ts
@@ -1,0 +1,4 @@
+import Vue from "vue";
+
+export class DwfDmsInput extends Vue {
+}
